Completed the submit transaction api, tested the apis using postman

+++ src/main/java/com/transfer/application/dtos/transactions/SubmitTransaction.java
@@ -10,10 +10,10 @@ import lombok.Data;
 public class SubmitTransaction {
 
     @Positive
-    private Integer sourceAccountId;
+    private Long sourceAccountId;
 
     @Positive
-    private Integer destinationAccountId;
+    private Long destinationAccountId;
 
     @Positive
     private Double amount;
